box: rewrite rollback to savepoint to Lua/C
The fix is prevent an emerge of "FFI sandwich".
See https://github.com/tarantool/tarantool/issues/4427#issuecomment-524798109
for more information.

One of the aims of this change (in addition to the fix) is to increase locality of highly
coupled code for the rollback to savepoint.

Noted that <struct txn>.id starts from 1, because I lean on this fact to
use luaL_toint64(), which does not distinguish an unexpected Lua type
case and cdata<int64_t> with zero value. It seems that this assumption
already exists: the code that prepare arguments for 'on_commit' triggers
uses luaL_toint64() too (see lbox_txn_pairs()).

Fixes #4427

Co-authored-by: Alexander Turenko<alexander.turenko@tarantool.org>
Co-authored-by: Leonid Vasiliev<lvasiliev@tarantool.org>