sql: use rowid as PK of ephemeral space on INSERT
Prior to this patch, the primary key of the ephemeral space always
consisted of all columns of the ephemeral space. This can lead to
an error during insertion when using ephemeral space. The problem
was that after selecting from the ephemeral space, the values were
sorted differently than when they were initially inserted.

This patch fixes sorting in ephemeral space in case it was used in
INSERT.

Closes #4256